pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "Python MCP Server template"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Alpic" }]
keywords = ["mcp", "llm", "alpic", "python"]
dependencies = [
    "black>=25.1.0",
    "mcp",
    "python-dotenv>=1.1.1",
    "zeroentropy>=0.1.0a6",
]
